+++ internal/scraper/scraper.go
@@ -2,110 +2,60 @@ package scraper
 
 import (
 	"fmt"
-	"io"
-	"net/http"
-	"strings"
-
-	"golang.org/x/net/html"
 )
 
-type Set map[string]struct{}
+// ScrapeWebsite recursively scrapes a website up to a maximum depth (maxIterations).
+// It starts with an initial set of URLs (currentPage) and maintains a set of all
+// discovered links (allLinks).
+func ScrapeWebsite(currentPage, allLinks Set, iteration int32, urlPrefix string) Set {
+	const maxIterations = 50
 
-type Broken map[string]int
-
-func NewSet() Set {
-	return make(Set)
-}
-
-func (s Set) Add(val string) {
-	s[val] = struct{}{}
-}
-
-func (s Set) Count() int {
-	return len(s)
-}
-
-func HtmlParser(r io.Reader) (Set, error) {
-	doc, err := html.Parse(r)
-	if err != nil {
-		return nil, err
+	// Base Cases:
+	// 1. Stop if we've reached the maximum iteration depth.
+	if iteration > maxIterations {
+		return allLinks
 	}
 
-	results := NewSet()
-	for n := range doc.Descendants() {
-		if n.Type == html.ElementNode && n.Data == "a" {
-			for _, attr := range n.Attr {
-				if attr.Key == "href" {
-					if IsValidUrl(attr.Val) {
-						results.Add(attr.Val)
-					}
-				}
-			}
+	// 2. Stop if no new links were found in the previous iteration.
+	initialLinkCount := allLinks.Count()
+	for link := range currentPage {
+		allLinks.Add(link)
+	}
+	if initialLinkCount == allLinks.Count() && iteration > 0 {
+		return allLinks
+	}
+
+	// Recursive Step:
+	// Explore links found on the current page.
+	allLinks = scrapePageLinks(currentPage, urlPrefix, allLinks, iteration)
+
+	return allLinks
+}
+
+// scrapePageLinks iterates through a set of URLs (currentPage), fetches and parses
+// the HTML content of each, and recursively calls ScrapeWebsite to explore any
+// new links found.
+func scrapePageLinks(currentPage Set, urlPrefix string, allLinks Set, iteration int32) Set {
+	for url := range currentPage {
+		if !IsValidUrl(url) {
+			continue // Skip invalid URLs
 		}
-	}
 
-	if len(results) > 0 {
-		return results, nil
-	}
-
-	return nil, nil
-}
-
-func IsValidUrl(url string) bool {
-	if url == "#" {
-		return false
-	}
-	if strings.HasPrefix(url, "http") {
-		return false
-	}
-	if strings.HasPrefix(url, "/") {
-		return true
-	}
-	return false
-}
-
-func GetHtml(requestURL string, REQ_URL_PREFIX string) (io.ReadCloser, error) {
-	res, err := http.Get(REQ_URL_PREFIX + requestURL)
-	if err != nil {
-		return nil, fmt.Errorf("error making http request: %w", err) // Wrap the error
-	}
-	return res.Body, nil
-}
-
-func ScrapeWebsite(current_page Set, all_links Set, iter int32, REQ_URL_PREFIX string) Set {
-	if iter > 50 {
-		return all_links
-	}
-	current_count := all_links.Count()
-	if all_links.Count() != 0 {
-		for k := range current_page {
-			all_links.Add(k)
+		htmlContent, err := GetHtml(url, urlPrefix)
+		if err != nil {
+			fmt.Println("Error fetching URL:", err)
+			continue // Skip URLs that can't be fetched
 		}
-		if current_count == all_links.Count() {
-			return all_links
+
+		newLinks, err := HtmlParser(htmlContent)
+		if err != nil {
+			fmt.Println("Error parsing HTML:", err)
+			continue // Skip pages that can't be parsed
 		}
-	} else {
-		for k := range current_page {
-			all_links.Add(k)
-		}
+
+		// Recursively scrape the newly found links.
+		allLinks = ScrapeWebsite(newLinks, allLinks, iteration+1, urlPrefix)
 	}
 
-	for k := range current_page {
-		if IsValidUrl(k) {
-			body, err := GetHtml(k, REQ_URL_PREFIX)
-			if err != nil {
-				fmt.Println(err)
-				continue
-			}
-			routes, err := HtmlParser(body)
-			if err != nil {
-				fmt.Println(err) // Handle the error from HtmlParser
-				continue         //Continue with the loop
-			}
-
-			all_links = ScrapeWebsite(routes, all_links, iter+1, REQ_URL_PREFIX)
-		}
-	}
-
-	return all_links
+	return allLinks
 }
